# docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: ComfyUI Serverless API
  version: 1.0.0
  description: |
    Serverless ComfyUI API for AI image generation and workflow execution.
    
    This API provides a complete solution for running ComfyUI workflows in a serverless environment,
    with features including:
    - Asynchronous workflow execution
    - Model management and caching  
    - File upload/download
    - Real-time progress tracking
    - Webhook notifications
    - Authentication and rate limiting
    
  contact:
    name: API Support
    email: support@comfyui-serverless.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.comfyui-serverless.com
    description: Production server
  - url: https://staging-api.comfyui-serverless.com
    description: Staging server
  - url: http://localhost:8000
    description: Local development server
security:
  - BearerAuth: []
paths:
...
  /auth/register:
    post:
      summary: Register User
      description: Register a new user account
      tags:
        - Authentication
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        '200':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Registration failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/login:
    post:
      summary: User Login
      description: Authenticate user and return access tokens
      tags:
        - Authentication
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: "user@example.com"
                password:
                  type: string
                  example: "securepassword"
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Token'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/refresh:
    post:
      summary: Refresh Token
      description: Refresh access token using refresh token
      tags:
        - Authentication
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refresh_token
              properties:
                refresh_token:
                  type: string
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Token'
        '401':
          description: Invalid refresh token
  /auth/me:
    get:
      summary: Get Current User
      description: Get current user information
      tags:
        - Authentication
      responses:
        '200':
          description: User information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Unauthorized
...
  /workflows/:
    get:
      summary: List Workflow Executions
      description: List user's workflow executions with filtering
      tags:
        - Workflows
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/WorkflowStatus'
          description: Filter by status
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Number of results
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Offset for pagination
      responses:
        '200':
          description: List of workflow executions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WorkflowResult'
...
  /models/:
    get:
      summary: List Models
      description: List all available models with their status
      tags:
        - Models
      parameters:
        - name: model_type
          in: query
          schema:
            $ref: '#/components/schemas/ModelType'
          description: Filter by model type
        - name: available_only
          in: query
          schema:
            type: boolean
            default: false
          description: Show only available models
      responses:
        '200':
          description: List of models
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelListResponse'
...
  /files/upload:
    post:
      summary: Upload File
      description: Upload a file for use in workflows
      tags:
        - Files
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                description:
                  type: string
      responses:
        '200':
          description: File uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileUploadResponse'
        '413':
          description: File too large
  /files/{file_id}:
    get:
      summary: Download File
      description: Download a file by ID
      tags:
        - Files
      parameters:
        - name: file_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: File content
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '404':
          description: File not found
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserCreate:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: "user@example.com"
        username:
          type: string
          example: "johndoe"
        password:
          type: string
          minLength: 8
          example: "securepassword123"
        is_active:
          type: boolean
          default: true
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        email:
          type: string
          format: email
        username:
          type: string
          nullable: true
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
          nullable: true
    Token:
      type: object
      properties:
        access_token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        refresh_token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        token_type:
          type: string
          example: "bearer"
        expires_in:
          type: integer
          example: 1800

    # Workflow schemas
    WorkflowStatus:
      type: string
      enum:
        - pending
        - running
        - completed
        - failed
        - cancelled
      example: "pending"
...
    WorkflowResult:
      type: object
      required:
        - execution_id
        - status
      properties:
        execution_id:
          type: string
        status:
          $ref: '#/components/schemas/WorkflowStatus'
        started_at:
          type: string
          format: date-time
          nullable: true
        completed_at:
          type: string
          format: date-time
          nullable: true
        duration_seconds:
          type: number
          format: float
          nullable: true
        outputs:
          type: object
          additionalProperties: true
          nullable: true
        error:
          type: string
          nullable: true
        logs:
          type: array
          items:
            type: object
          nullable: true
        metadata:
          type: object
          additionalProperties: true
          nullable: true

    # Model schemas
    ModelType:
      type: string
      enum:
        - checkpoint
        - lora
        - embedding
        - vae
        - controlnet
        - upscaler
      example: "checkpoint"
    ModelStatus:
      type: object
      required:
        - name
        - type
        - is_loaded
        - is_downloading
      properties:
        name:
          type: string
          example: "sd-v1-5-inpainting.ckpt"
        type:
          $ref: '#/components/schemas/ModelType'
        is_loaded:
          type: boolean
        is_downloading:
          type: boolean
        download_progress:
          type: number
          format: float
          minimum: 0.0
          maximum: 1.0
          nullable: true
        last_used:
          type: string
          format: date-time
          nullable: true
        memory_usage_mb:
          type: number
          format: float
          nullable: true
    ModelListResponse:
      type: object
      required:
        - models
        - total_memory_usage_mb
        - available_memory_mb
      properties:
        models:
          type: array
          items:
            $ref: '#/components/schemas/ModelStatus'
        total_memory_usage_mb:
          type: number
          format: float
        available_memory_mb:
          type: number
          format: float

    # File schemas
    FileUploadResponse:
      type: object
      required:
        - file_id
        - filename
        - size
        - content_type
      properties:
        file_id:
          type: string
          example: "file_abc123"
        filename:
          type: string
          example: "input_image.png"
        size:
          type: integer
          example: 1048576
        content_type:
          type: string
          example: "image/png"
        upload_url:
          type: string
          format: uri
          nullable: true
...
    ErrorDetail:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: "VALIDATION_ERROR"
        message:
          type: string
          example: "Field 'workflow.nodes' is required"
        field:
          type: string
          nullable: true
          example: "workflow.nodes"
    ErrorResponse:
      type: object
      required:
        - error
        - message
        - timestamp
      properties:
        error:
          type: string
          example: "validation_error"
        message:
          type: string
          example: "Invalid request data"
        details:
          type: array
          items:
            $ref: '#/components/schemas/ErrorDetail'
          nullable: true
        timestamp:
          type: string
          format: date-time
        request_id:
          type: string
          nullable: true
          example: "req_abc123"
...